código QUASE pronto (cheio de erro)

+++ Util.h
@@ -5,7 +5,7 @@ void* mallocSafe(size_t nbytes);
 
 void verificaArquivo(FILE* arquivo);
 
-int calculaDeslocamentoPagina(int posPagina, size_t tamanho_registro);
+size_t calculaDeslocamentoPagina(int posPagina, size_t tamanho_registro);
 
 void limpaBuffer();
